+++ openquake/calculators/classical_damage.py
@@ -40,7 +40,7 @@ def classical_damage(riskinputs, crmodel, param, monitor):
     """
     result = AccumDict(accum=AccumDict())
     for ri in riskinputs:
-        for out in crmodel.gen_outputs(ri, monitor):
+        for out in ri.gen_outputs(crmodel, monitor):
             for l, loss_type in enumerate(crmodel.loss_types):
                 ordinals = ri.assets['ordinal']
                 result[l, out.rlzi] += dict(zip(ordinals, out[loss_type]))
