+++ openquake/calculators/ebrisk.py
@@ -23,8 +23,7 @@ from openquake.baselib import hdf5, datastore, parallel, general
 from openquake.baselib.python3compat import zip, encode
 from openquake.hazardlib.stats import set_rlzs_stats
 from openquake.risklib.scientific import losses_by_period, LossesByAsset
-from openquake.risklib.riskmodels import get_assets_by_taxo
-from openquake.risklib.riskinput import cache_epsilons
+from openquake.risklib.riskinput import cache_epsilons, get_assets_by_taxo
 from openquake.calculators import base, event_based, getters
 from openquake.calculators.export.loss_curves import get_loss_builder
